requirements.txt
pyqt6>=6.10.0
openai>=2.7.0
python-dotenv>=1.2.0
newspaper3k>=0.2.8
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.27.0
lxml_html_clean>=0.4.0
requests>=2.32.0
beautifulsoup4>=4.14.0
